avoid conflicting with oF hoisted namespaces
When compiling on the Raspberry Pi / Wheezy Raspbian, ofTypes hoists shared_ptr, weak_ptr and enable_shared_from_this from std::tr1 while this code was hosting them just from std.

__cplusplus seems to be defined to 1 when compiling C++ and for more recently compilers, will be the value of the compliance for example 199711 by default or 201103 for c++11. Compiling on the Raspberry Pi of course evaluates to 1.

I cannot be certain that this is correct, but the change does allow complication on the Raspberry Pi so I'm good with that until a better solution is found.